refactor k8s communication
